modified TestLambaUtilities in order to pass al the gradlew build tests

+++ 101-lambda-utilities-for-lists-and-maps/src/test/java/it/unibo/oop/lab/lambda/TestLambdaUtilities.java
@@ -88,9 +88,9 @@ final class TestLambdaUtilities {
             () -> random.nextInt(10) - 10
         );
         assertEquals(map.keySet(), Set.of("p1", "p2", "p3", "n1", "n2"));
-        assertEquals(map.get("p1").intValue(), 1);
-        assertEquals(map.get("p2").intValue(), 2);
-        assertEquals(map.get("p3").intValue(), 3);
+        assertEquals(map.get("p1"), 1);
+        assertEquals(map.get("p2"), 2);
+        assertEquals(map.get("p3"), 3);
         assertTrue(map.get("n1") < 0);
         assertTrue(map.get("n2") < 0);
     }
